[ui-components] refactor: moved erroricon/searchicon
