+++ packages/editor-common/web/src/Components/LinkComponents/LinkPanel.jsx
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { mergeStyles, isValidUrl } from 'wix-rich-content-common';
 import Tooltip from 'wix-rich-content-common/libs/Tooltip';
-import { Checkbox } from 'wix-rich-content-ui-components';
-import { ErrorIcon } from '../../Icons';
+import { Checkbox, ErrorIcon } from 'wix-rich-content-ui-components';
 import styles from '../../../statics/styles/link-panel.scss';
 import { LinkPanelDropdown } from './LinkPanelDropdown';
