+++ packages/plugin-giphy/web/src/icons/index.ts
@@ -1,6 +1,5 @@
 export { default as InsertPluginIcon } from './InsertPluginIcon';
 export { default as MediaReplaceIcon } from './MediaReplaceIcon';
-export { default as SearchIcon } from './SearchIcon';
 export { default as InsertPluginMobileIcon } from './InsertPluginMobileIcon';
 export { default as CloseIcon } from './CloseIcon';
 export { default as PoweredByGiphy } from './PoweredByGiphy';
