+++ packages/ui-components/web/src/Components/TextInput.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { mergeStyles } from 'wix-rich-content-common';
 import Tooltip from 'wix-rich-content-common/libs/Tooltip';
-import { ErrorIcon, SearchIcon } from 'wix-rich-content-editor-common';
+import { ErrorIcon, SearchIcon } from '../Icons';
 
 import textInputStyles from '../../statics/styles/text-input.scss';
 import { omit } from 'lodash';
