+++ packages/ui-components/web/src/Icons/index.ts
@@ -30,3 +30,4 @@ export { default as ClearIcon } from './ClearIcon';
 export { default as SearchIcon } from './SearchIcon';
 export { default as DropdownArrowIcon } from './DropdownArrowIcon';
 export { default as CheckIcon } from './CheckIcon';
+export { default as ErrorIcon } from './ErrorIcon';
